review comments, consolidate stack implementations
